Split native Zenmap XML file action helpers
